backend/requirements.txt
mysql~=0.0.3
mysql-connector-python~=8.0.33
fastapi~=0.103.0
pydantic~=2.3.0
passlib~=1.7.4
python-jose~=3.3.0